+++ Kernel/Modules/AgentTicketCompose.pm
@@ -1792,6 +1792,9 @@ sub Run {
         # run compose modules
         if ( ref $ConfigObject->Get('Ticket::Frontend::ArticleComposeModule') eq 'HASH' ) {
 
+            # use ticket QueueID in compose modules
+            $GetParam{QueueID} = $Ticket{QueueID};
+
             my %Jobs = %{ $ConfigObject->Get('Ticket::Frontend::ArticleComposeModule') };
             for my $Job ( sort keys %Jobs ) {
 
@@ -1813,7 +1816,13 @@ sub Run {
                 }
 
                 # run module
-                $Object->Run( %GetParam, Config => $Jobs{$Job} );
+                my $NewParams = $Object->Run( %GetParam, Config => $Jobs{$Job} );
+
+                if ($NewParams) {
+                    for my $Parameter ( $Object->Option( %GetParam, Config => $Jobs{$Job} ) ) {
+                        $GetParam{$Parameter} = $NewParams;
+                    }
+                }
 
                 # get errors
                 %Error = (